Refactored performance critical code
